Split QueryStoreService.cs into partial classes
Move-only refactor; no behavior changes. QueryStoreService.cs (1,403
lines) split into 3 partials:

  Grouped    (584) - FetchGroupedByQueryHashAsync +
                     FetchGroupedByModuleAsync
  WaitStats  (232) - IsWaitStatsCaptureEnabledAsync +
                     FetchGlobalWaitStatsAsync +
                     FetchPlanWaitStatsAsync +
                     FetchGlobalWaitStatsRibbonAsync +
                     BuildWaitProfile

Main file now 584 lines — top-of-pipe fetches: CheckEnabledAsync,
FetchTopPlansAsync, FetchAggregateHistoryAsync, FetchTimeSliceDataAsync.

Class made `static partial`.

Build clean: 0 errors, 0 warnings on PlanViewer.Core.
